전반적인 리팩토리실시 ( 20.12.02 )
함수명 변경 , 이벤트처리 메서드 추출실시 ,
파일분리 정리완료
